# studyai-news-api/openapi/news-api.yaml
openapi: 3.1.0
info:
  title: StudyAI News API
  version: 0.5.0
  description: Versioned contract for public News APIs and editorial APIs reached only through the authenticated StudyAINow admin proxy.
servers:
  - url: https://news.studyai.now
paths:
...
  /api/admin/news/sources:
    get:
      operationId: listNewsSources
      summary: List approved source configuration and ingestion health
      security:
        - ingestionOperator: []
      responses:
        "200":
          description: Source health list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceHealthResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createNewsSource
      summary: Create a paused source pending policy review
      security:
        - ingestionOperator: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SourceCreateRequest"
      responses:
        "201":
          description: Paused source created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceMutationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/admin/news/sources/probe:
    post:
      operationId: probeNewsSource
      summary: Validate robots policy, safety limits, feed structure and quality
      security:
        - ingestionOperator: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SourceProbeRequest"
      responses:
        "200":
          description: Source probe result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceProbeResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/admin/news/sources/{sourceId}:
    parameters:
      - $ref: "#/components/parameters/SourceId"
    patch:
      operationId: updateNewsSource
      summary: Update source status or policy review state
      security:
        - ingestionOperator: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SourceUpdateRequest"
      responses:
        "200":
          description: Source updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceMutationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: retireNewsSource
      summary: Soft-retire a source without deleting evidence
      security:
        - ingestionOperator: []
      responses:
        "200":
          description: Source retired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceMutationResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/admin/news/sources/{sourceId}/run:
    parameters:
      - $ref: "#/components/parameters/SourceId"
    post:
      operationId: runNewsSource
      summary: Run one approved source with idempotency protection
      security:
        - ingestionOperator: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            minLength: 8
            maxLength: 120
      responses:
        "200":
          description: Ingestion completed or was deduplicated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManualIngestionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "502":
          description: Source ingestion failed and was recorded for retry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManualIngestionResponse"
  /api/admin/news/dashboard:
    get:
      operationId: getNewsEditorialDashboard
      summary: Read editorial queue counts
      security:
        - ingestionOperator: []
        - studyAiAdminService: []
      responses:
        "200":
          $ref: "#/components/responses/AdminData"
  /api/admin/news/candidates:
    get:
      operationId: listNewsCandidates
      summary: List clustered and enriched source candidates
      security:
        - ingestionOperator: []
        - studyAiAdminService: []
      responses:
        "200":
          $ref: "#/components/responses/AdminData"
  /api/admin/news/candidates/enrich:
    post:
      operationId: enrichNewsCandidates
      summary: Run deterministic deduplication and controlled metadata enrichment
      security:
        - ingestionOperator: []
        - studyAiAdminService: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          $ref: "#/components/responses/AdminData"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/admin/news/candidates/{storyId}:
    parameters:
      - $ref: "#/components/parameters/StoryId"
    patch:
      operationId: updateNewsCandidateMetadata
      summary: Human-review and lock a candidate category and tags
      security:
        - ingestionOperator: []
        - studyAiAdminService: []
      responses:
        "200":
          $ref: "#/components/responses/AdminData"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/admin/news/stories/{storyId}:
    parameters:
      - $ref: "#/components/parameters/StoryId"
    get:
      operationId: getNewsStoryResearch
      summary: Read a story research package, source evidence and Claim Ledger
      security:
        - ingestionOperator: []
        - studyAiAdminService: []
      responses:
        "200":
          $ref: "#/components/responses/AdminData"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/admin/news/stories/{storyId}/research:
    parameters:
      - $ref: "#/components/parameters/StoryId"
    post:
      operationId: generateNewsStoryResearch
      summary: Generate or reuse an immutable deterministic source-bound research package
      security:
        - ingestionOperator: []
        - studyAiAdminService: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          $ref: "#/components/responses/AdminData"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/admin/news/stories/{storyId}/claims:
    parameters:
      - $ref: "#/components/parameters/StoryId"
    post:
      operationId: createNewsClaim
      summary: Add an explicitly typed human Claim pending source evidence
      security:
        - ingestionOperator: []
        - studyAiAdminService: []
      responses:
        "201":
          $ref: "#/components/responses/AdminData"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/admin/news/claims/{claimId}:
    patch:
      operationId: reviewNewsClaim
      summary: Review a Claim and append its immutable revision record
      security:
        - ingestionOperator: []
        - studyAiAdminService: []
      responses:
        "200":
          $ref: "#/components/responses/AdminData"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/admin/news/claims/{claimId}/evidence:
    post:
      operationId: addNewsClaimEvidence
      summary: Attach immutable evidence from a source item in the same story
      security:
        - ingestionOperator: []
        - studyAiAdminService: []
      responses:
        "201":
          $ref: "#/components/responses/AdminData"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/admin/news/articles:
    get:
      operationId: listNewsArticles
      summary: List editorial articles and publication states
      security:
        - ingestionOperator: []
        - studyAiAdminService: []
      responses:
        "200":
          $ref: "#/components/responses/AdminData"
    post:
      operationId: createNewsArticle
      summary: Create a manually edited article and immutable first revision
      security:
        - ingestionOperator: []
        - studyAiAdminService: []
      responses:
        "201":
          $ref: "#/components/responses/AdminData"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/admin/news/articles/{articleId}:
    parameters:
      - $ref: "#/components/parameters/ArticleId"
    get:
      operationId: getNewsArticle
      summary: Read an article with revisions, approvals, taxonomy and audit history
      security:
        - ingestionOperator: []
        - studyAiAdminService: []
      responses:
        "200":
          $ref: "#/components/responses/AdminData"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: reviseNewsArticle
      summary: Save an immutable article revision using optimistic concurrency
      security:
        - ingestionOperator: []
        - studyAiAdminService: []
      responses:
        "200":
          $ref: "#/components/responses/AdminData"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: Article version conflict
  /api/admin/news/articles/{articleId}/actions/{action}:
    parameters:
      - $ref: "#/components/parameters/ArticleId"
      - name: action
        in: path
        required: true
        schema:
          type: string
          enum: [submit, return, reject, approve, schedule, publish, correct, withdraw, reopen]
    post:
      operationId: performNewsArticleAction
      summary: Perform an audited, idempotent editorial state transition
      security:
        - ingestionOperator: []
        - studyAiAdminService: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          $ref: "#/components/responses/AdminData"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/admin/news/taxonomy:
    get:
      operationId: listNewsTaxonomy
      summary: List controlled primary categories and tags
      security:
        - ingestionOperator: []
        - studyAiAdminService: []
      responses:
        "200":
          $ref: "#/components/responses/AdminData"
  /api/admin/news/taxonomy/tags:
    post:
      operationId: createNewsTag
      summary: Create a human-controlled tag
      security:
        - ingestionOperator: []
        - studyAiAdminService: []
      responses:
        "201":
          $ref: "#/components/responses/AdminData"
  /api/admin/news/taxonomy/{taxonomyId}:
    parameters:
      - $ref: "#/components/parameters/TaxonomyId"
    patch:
      operationId: updateNewsTaxonomy
      summary: Update a category or tag display name, aliases and status
      security:
        - ingestionOperator: []
        - studyAiAdminService: []
      responses:
        "200":
          $ref: "#/components/responses/AdminData"
  /api/admin/news/taxonomy/{taxonomyId}/merge:
    parameters:
      - $ref: "#/components/parameters/TaxonomyId"
    post:
      operationId: mergeNewsTag
      summary: Merge a tag into an active canonical tag
      security:
        - ingestionOperator: []
        - studyAiAdminService: []
      responses:
        "200":
          $ref: "#/components/responses/AdminData"
components:
  securitySchemes:
    ingestionOperator:
      type: http
      scheme: bearer
      bearerFormat: opaque-operator-token
    studyAiAdminService:
      type: apiKey
      in: header
      name: X-StudyAI-Admin-Service-Token
      description: Internal credential sent only by the authenticated studyai.now admin proxy.
  parameters:
    SourceId:
      name: sourceId
      in: path
      required: true
      schema:
        type: string
        pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
...
  schemas:
...
    SourceHealthResponse:
      type: object
      additionalProperties: false
      required: [ok, sources, traceId]
      properties:
        ok:
          type: boolean
          const: true
        sources:
          type: array
          items:
            $ref: "#/components/schemas/SourceHealth"
        traceId:
          type: string
...
    SourceMutationResponse:
      type: object
      additionalProperties: true
      required: [ok, traceId]
      properties:
        ok:
          type: boolean
          const: true
        sourceId:
          type: string
        status:
          type: string
        policyStatus:
          type: string
        traceId:
          type: string
...
  responses:
...
    Unauthorized:
      description: StudyAINow admin service identity or operational credential is missing or invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...